+++ lib/main_pages/edit_profile_screen.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/material.dart';
+import 'package:learnvironment/data/notification_storage.dart';
 import 'package:learnvironment/data/user_data.dart';
 import 'package:learnvironment/main_pages/profile_screen.dart';
 import 'package:learnvironment/services/data_service.dart';
-import 'package:learnvironment/services/auth_service.dart';
+import 'package:learnvironment/services/firebase/auth_service.dart';
 import 'package:provider/provider.dart';
 import 'package:http/http.dart' as http;
 
@@ -113,6 +114,11 @@ class EditProfilePageState extends State<EditProfilePage> {
       if (userData == null) {
         _showErrorDialog("No user is logged in.", "Error");
         return;
+      }
+
+      if (accountType != userData.role) {
+        NotificationStorage.notificationMessages = [];
+        authService.fetchedNotifications = true;
       }
 
       if (email != userData.email) {