+++ lib/teacher/teacher_home.dart
@@ -2,7 +2,6 @@ import 'package:flutter/material.dart';
 import 'package:learnvironment/teacher/statistics_teacher_page.dart';
 import 'package:learnvironment/teacher/teacher_main_page.dart';
 import 'package:learnvironment/main_pages/profile_screen.dart';
-import 'package:learnvironment/main_pages/statistics_page.dart';
 import 'package:learnvironment/teacher/games_page_teacher.dart';
 import 'package:learnvironment/main_pages/notifications_page.dart';