+++ test/developer/CreateGames/create_drag_test.dart
@@ -6,7 +6,7 @@ import 'package:learnvironment/data/game_data.dart';
 import 'package:learnvironment/developer/CreateGames/create_drag.dart';
 import 'package:learnvironment/developer/new_game.dart';
 import 'package:mockito/mockito.dart';
-import 'package:learnvironment/services/auth_service.dart';
+import 'package:learnvironment/services/firebase/auth_service.dart';
 import 'package:learnvironment/services/data_service.dart';
 import 'package:provider/provider.dart';