+++ test/games_templates/results_page_test.dart
@@ -5,7 +5,7 @@ import 'package:learnvironment/games_templates/results_page.dart';
 import 'package:mockito/mockito.dart';
 import 'package:learnvironment/services/data_service.dart';
 import 'package:learnvironment/data/user_data.dart';
-import 'package:learnvironment/services/auth_service.dart';
+import 'package:learnvironment/services/firebase/auth_service.dart';
 import 'package:firebase_auth_mocks/firebase_auth_mocks.dart';
 import 'package:provider/provider.dart';
 
@@ -30,7 +30,6 @@ class MockDataService extends Mock implements DataService {
 }
 
 void main() {
-
   group('ResultsPage Tests', ()
   {
     late Widget testWidget;